requirements.txt
wheel
setuptools~=41.2.0
twine
pytest
pytest-runner
matplotlib==3.2.2
backtrader==1.9.76.123
get_all_tickers
tda-api==1.1.4
webdriver-manager
beautifulsoup4
git+https://github.com/mariostoev/finviz
selenium
httpx~=0.16.1
pandas~=1.2.2
numpy~=1.20.1